Extract subject macros into their own module
