+++ lib/psulib_traject.rb
@@ -17,6 +17,7 @@ module PsulibTraject
   require 'psulib_traject/hathi_overlap_reducer'
   require 'psulib_traject/holdings'
   require 'psulib_traject/macros'
+  require 'psulib_traject/macros/subjects'
   require 'psulib_traject/marc_combining_reader'
   require 'psulib_traject/null_object'
   require 'psulib_traject/processors/access_facet'
